+++ src/main/java/com/tiyler/roadstransport/model/MailArea.java
@@ -2,5 +2,5 @@ package com.tiyler.roadstransport.model;
 
 import java.util.UUID;
 
-public record MailArea(UUID ownerId, BlockKey center) {
+public record MailArea(UUID ownerId, BlockKey center, String name, BlockKey signBlock) {
 }